requirements.txt
blinker==1.7.0
click==8.1.7
colored==2.2.4
dnspython==2.6.1
feedparser==6.0.11
Flask==3.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
pymongo==4.6.3
PyYAML==6.0.1
regex==2023.12.25
sgmllib3k==1.0.0
tqdm==4.66.2
Werkzeug==3.0.1
